Save,Update,deleteControllers

+++ Shopy/src/main/java/Ecommerce/example/Shopy/Repository/Product_PhotoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface Product_PhotoRepository extends JpaRepository<Product_Photo, Integer> {
+public interface Product_PhotoRepository extends JpaRepository<Product_Photo, Long> {
 }
